% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGeometeR-package.R
\docType{package}
\name{compGeometeR-package}
\alias{compGeometeR}
\alias{compGeometeR-package}
\title{compGeometeR: Computational Geometry Algorithms}
\description{
Computational Geometry Algorithms for use in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/manaakiwhenua/compGeometeR}
  \item Report bugs at \url{https://github.com/manaakiwhenua/compGeometeR/issues}
}

}
\author{
\strong{Maintainer}: O. Pascal Omondiagbe \email{omondiagbep@landcareresearch.co.nz}

Authors:
\itemize{
  \item Thomas R. Etherington \email{etheringtont@landcareresearch.co.nz}
}

}
\keyword{internal}
