% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGeometeR.R
\docType{package}
\name{compGeometeR}
\alias{compGeometeR}
\title{Computational geometry algorithms for R}
\description{
Implementation of various computation geomtery algorithms for use in R

\tabular{ll}{ Package: \tab compGeometeR\cr  Date: \tab
2020-05-20\cr License: \tab GPL-2\cr LazyLoad: \tab yes\cr }
}
\references{
http://www.qhull.org/html/qh-code.htm
}
\author{
Pascal Omondiagbe\cr Tom Etherington\cr Maintainers: pascal Omondiagbe <omondiagbep@landcareresearch.co.nz>
}
\keyword{package}
