% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex-hull.R
\name{convex_hull}
\alias{convex_hull}
\title{Convex hull}
\usage{
convex_hull(points = NULL)
}
\arguments{
\item{points}{a \eqn{n}-by-\eqn{d} dataframe or matrix. The rows
represent \eqn{n} points and the \eqn{d} columns the coordinates in 
\eqn{d}-dimensional space.}
}
\value{
Returns a list consisting of:

\itemize{
  \item \code{input_points}: the input points used to create the convex hull.
  \item \code{hull_simplices}: a \eqn{s}-by-\eqn{d} matrix of point indices 
  that define the \eqn{s} \href{https://en.wikipedia.org/wiki/Simplex}{simplices} 
  that make up the convex hull.
  \item \code{hull_indices}: a vector of the point indices that form the 
  convex hull.
  \item \code{hull_vertices}: a matrix of point coordinates that form the 
  convex hull.
}

In the \eqn{2}-dimensional case the convex hull indices and vertices are 
returned in a circular order to ease plotting, but in other dimensions there 
is no specific order.
}
\description{
This function calculates the 
\href{https://en.wikipedia.org/wiki/Convex_hull}{convex hull} around a set of
\eqn{n} points in \eqn{d}-dimensional space using the
\href{http://www.qhull.org}{Qhull} library.
}
\examples{
# Define points
x <- c(30, 70, 20, 50, 40, 70)
y <- c(35, 80, 70, 50, 60, 20)
p <- data.frame(x, y)
# Create convex hull and plot
ch <- convex_hull(points=p)
plot(p, pch = as.character(seq(nrow(p))))
polygon(ch$hull_vertices, border="red")

}
\references{
Barber CB, Dobkin DP, Huhdanpaa H (1996) The Quickhull algorithm 
for convex hulls. ACM Transactions on Mathematical Software, 22(4):469-83 
\url{https://doi.org/10.1145/235815.235821}.
}
\seealso{
\code{\link{convex_layer}}
}
