% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displace-coordinates.R
\name{displace_coordinates}
\alias{displace_coordinates}
\title{Displace coordinates}
\usage{
displace_coordinates(points = NULL, radii)
}
\arguments{
\item{points}{a \eqn{n}-by-\eqn{d} dataframe or matrix. The rows
represent \eqn{n} points and the \eqn{d} columns the coordinates in 
\eqn{d}-dimensional space.}

\item{radii}{a vector of real numbers that defines the radius of the 
{eqn{n}-ball} within which each point will be displaced.}
}
\value{
Returns a \eqn{n}-by-\eqn{d} dataframe of coordinates that have been 
  displaced based on the specified radii.
}
\description{
This function displaces each point coordinate within an 
  \href{https://en.wikipedia.org/wiki/Ball_(mathematics)}{eqn{n}-ball} of 
  specified radius.
}
\examples{
# Define points
x <- c(30, 70, 20, 50, 40, 70)
y <- c(35, 80, 70, 50, 60, 20)
p <- data.frame(x, y)
# Define uncertainty radii
r <- c(2, 7, 5, 12, 0, 15)
# Plot points with uncertainty radii
plot(p, pch=16, asp=1)
symbols(x=p[,1], y=p[,2], circles = p[,3], fg="blue", inches = FALSE, add = TRUE)
# Generate and plot some other possible locations
for (i in seq(10)) {
  u_boot = displace_coordinates(points = p, radii = r)
  points(u_boot, pch=16, col="red", cex=0.5)
}

}
