% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGeometeR-package.R
\docType{package}
\name{compGeometeR-package}
\alias{compGeometeR}
\alias{compGeometeR-package}
\title{compGeometeR: Computational Geometry Algorithms for use in R}
\description{
Package provides R bindings for the QHull library of computational geometry algorithms.
}
\author{
\strong{Maintainer}: O. Pascal Omondiagbe \email{omondiagbep@landcareresearch.co.nz}

Authors:
\itemize{
  \item Thomas R. Etherington \email{etheringtont@landcareresearch.co.nz}
}

}
\keyword{internal}
