% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displace-coordinates.R
\name{displace_coordinates}
\alias{displace_coordinates}
\title{Displace coordinates}
\usage{
displace_coordinates(points = NULL, displace)
}
\arguments{
\item{points}{a \eqn{n}-by-\eqn{d} dataframe or matrix. The rows
represent \eqn{n} points and the \eqn{d} columns the coordinates in 
\eqn{d}-dimensional space.}

\item{displace}{a vector of real numbers that defines the radius of the 
{\eqn{n}-ball} within which each point will be displaced.}
}
\value{
Returns a \eqn{n}-by-\eqn{d} dataframe of coordinates that have been 
  displaced based on the specified radii.
}
\description{
This function displaces each point coordinate within an 
  \href{https://en.wikipedia.org/wiki/Ball_(mathematics)}{\eqn{n}-ball} of 
  specified radius.
}
\examples{
# Define points
x <- c(30, 70, 20, 50, 40, 80)
y <- c(35, 80, 70, 50, 60, 20)
p <- data.frame(x, y)
# Define displacement radii
r <- c(25, 15, 10, 22, 5, 17)
# Plot points with displacement radii
cols <- c("red", "blue", "orange", "skyblue", "green3", "hotpink")
plot(p, pch=16, col=cols, asp=1, xlim=c(0,100), ylim=c(0,100))
symbols(x=p[,1], y=p[,2], circles = r, fg=cols, inches = FALSE, add = TRUE)
# Generate and plot some other possible locations
for (i in seq(100)) {
 u_boot <- displace_coordinates(points = p, displace = r)
 points(u_boot, pch=4, col=cols, cex=0.5)
}

}
