% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-coordinates.R
\name{grid_coordinates}
\alias{grid_coordinates}
\title{Grid Coordinates}
\usage{
grid_coordinates(mins, maxs, spacings)
}
\arguments{
\item{mins}{Vector of length \code{d} listing the grid coordinate minimum for 
each dimension.}

\item{maxs}{Vector of length \code{d} listing the grid coordinate maximum for 
each dimension.}

\item{spacings}{Vector of length \code{d} listing the grid coordinate spacing 
for each dimension.}
}
\value{
A list of two objects:

\itemize{
  \item A dataframe with \code{d} columns and a row for each grid coordinate 
  - so potentially lots of rows!
  \item A list of length \code{d} that contains the grid coordinates along 
  each dimension.
}
}
\description{
This function creates a \code{d}-dimensional grid of coordinates 
systematically spaced throughout the specified spatial extent.
}
\examples{
# Point space grid coordinates usage
grid = grid_coordinates(mins=c(0,0), maxs=c(10,15), spacings=c(1,1))

}
