% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-coordinates.R
\name{grid_coordinates}
\alias{grid_coordinates}
\title{Grid Coordinates}
\usage{
grid_coordinates(mins, maxs, spacings)
}
\arguments{
\item{mins}{Vector of length \code{n} listing the point space minimum for each
dimension.}

\item{maxs}{Vector of length \code{n} listing the point space maximum for each
dimension.}

\item{spacings}{Vector of length \code{n} listing the spacings between points 
in all dimensions.}
}
\value{
A list of two objects.  The first object is a dataframe with \code{n} 
columns and a row for each coordinate - so potentially lots of rows!  The 
second object is another list of length \code{n} that contains the 
coordinates along each dimension.
}
\description{
This function creates a \code{n}-dimensional grid of coordinates 
systematically spaced throughout the specified spatial extent.
}
\examples{
# Point space grid coordinates usage
grid = grid_coordinates(mins=c(0,0), maxs=c(10,15), spacings=c(1,1))

}
