% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-coordinates.R
\name{grid_coordinates}
\alias{grid_coordinates}
\title{Grid Coordinates}
\usage{
grid_coordinates(mins, maxs, nCoords)
}
\arguments{
\item{mins}{Vector of length \code{n} listing the point space minimum for each
dimension.}

\item{maxs}{Vector of length \code{n} listing the pointspace maximum for each
dimension.}

\item{nCoords}{Number of coordinates across the point space in all dimensions.}
}
\value{
A matrix with \code{n} columns.
}
\description{
Create an \code{n}-dimensional grid of coordinates across space.
}
\details{
This function creates a grid of coordinates systematically located
throughout the specified point space to enable visualisation of alpha shape .
The extent of the grid is given by the \code{mins} and \code{maxs}, and the number
of coordinates for each dimension is given by \code{nCoords}.
}
\examples{
# Point space grid coordinates usage
xy = grid_coordinates(mins=c(15,0), maxs=c(35,200), nCoords=5)
}
